Validate

+++ app/Http/Controllers/Admin/ProductController.php
@@ -41,9 +41,18 @@ class ProductController extends Controller
      */
     public function store(ProductRequest $request)
     {
-        $path = $request->file('image')->store('products');
+/*        $path = $request->file('image')->store('products');
         $params = $request->all();
         $params['image'] = $path;
+        Product::create($params);
+        return redirect()->route('products.index');*/
+        $params = $request->all();
+
+        unset($params['image']);
+        if ($request->has('image')) {
+            $params['image'] = $request->file('image')->store('products');
+        }
+
         Product::create($params);
         return redirect()->route('products.index');
     }
@@ -80,10 +89,17 @@ class ProductController extends Controller
      */
     public function update(ProductRequest $request, Product $product)
     {
-        Storage::delete($product->image);
+ /*       Storage::delete($product->image);
         $path = $request->file('image')->store('products');
         $params = $request->all();
         $params['image'] = $path;
+        */
+        $params = $request->all();
+        unset($params['image']);
+        if ($request->has('image')) {
+            Storage::delete($product->image);
+            $params['image'] = $request->file('image')->store('products');
+        }
         $product->update($params);
         return redirect()->route('products.index');
     }
